+++ lib/ios_gen.dart
@@ -70,7 +70,7 @@ class ObjectiveCCreate {
           .replaceAll(",", "");
 
       allContent += "${importStr}";
-      allContent += "${staticPropertyImplementation(value.properties)}";
+      allContent += "${staticPropertyImplementation(value)}";
       if (!_isStaticPropertiesOnly(value)) {
         allContent += getCustomClassImport(value);
         allContent += "\nNS_ASSUME_NONNULL_BEGIN\n";
@@ -151,6 +151,13 @@ class ObjectiveCCreate {
   static String method(List<MethodInfo> methods) {
     String result = "";
     methods.forEach((method) {
+      String methodComment =
+          "\n/// Dart method declaraction: ${method.originDeclaration}\n";
+      method.args.forEach((element) {
+        methodComment +=
+            "/// @param ${element.name} Agument ${element.name}, type: ${element.type}.\n";
+      });
+      result += methodComment;
       result += "- (" + getTypeString(method.returnType) + ")" + method.name;
       String argType = "";
       if (!method.args.isEmpty) {
@@ -172,12 +179,12 @@ class ObjectiveCCreate {
     return result;
   }
 
-  static String staticPropertyImplementation(List<Property> properties) {
+  static String staticPropertyImplementation(GenClassBean classBean) {
     String staticPropertyStr = "";
-    properties.forEach((property) {
+    classBean.properties.forEach((property) {
       if (property.isStatic) {
         staticPropertyStr +=
-            "#define ${prefix}${property.name}\t${defaultValueOf(property)}\n";
+            "#define ${prefix}${classBean.classInfo.name}_${property.name}\t${defaultValueOf(property)}\n";
       }
     });
     if (staticPropertyStr.isNotEmpty) {
@@ -315,6 +322,8 @@ class ObjectiveCCreate {
     Property property, {
     bool showNullTag = true,
   }) {
+    String propertyComment =
+        "\n/// Dart property declaration: ${property.originDeclaration.trim()}.\n";
     String propertyString = "@property (nonatomic, ";
     switch (typeOf(property)) {
       case ObjectivePropertType.base:
@@ -343,7 +352,7 @@ class ObjectiveCCreate {
         break;
       default:
     }
-    return propertyString;
+    return propertyComment + propertyString;
   }
 
   static String getSubTypeString(Property property,